requirements.txt
# Web框架
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# 图像处理和计算机视觉
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# 深度学习和AI模型
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0

# 数据处理和分析
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
autogluon.tabular>=0.8.0
xgboost>=1.7.0

# 可视化
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 向量数据库和RAG
langchain>=0.0.300
langchain-community>=0.0.30
faiss-cpu>=1.7.4

# 人脸识别
face-recognition>=1.3.0
dlib>=19.24.0


# API客户端
openai>=1.0.0
requests>=2.31.0

# 文件处理
python-multipart>=0.0.6

# 系统监控
psutil>=5.9.0

# 加密和安全
bcrypt>=4.0.0


# 可选：PaddleNLP (如果需要)
# paddlenlp>=2.6.0

# 可选：GPU支持
# torch-audio  # 如果需要音频处理
# torchaudio   # 如果需要音频处理
